+++ src/module_writer.rs
@@ -38,11 +38,18 @@ use tempfile::{tempdir, TempDir};
 use tracing::{debug, instrument};
 use zip::{self, DateTime, ZipWriter};
 
+/// A deterministic, arbitrary, non-zero timestamp that use used as `mtime`
+/// of headers when writing sdists.
+///
+/// This value, copied from the tar crate, corresponds to _Jul 23, 2006_,
+/// which is the date of the first commit for what would become Rust.
+///
+/// This value is used instead of unix epoch 0 because some tools do not handle
+/// the 0 value properly (See rust-lang/cargo#9512).
+const SDIST_DETERMINISTIC_TIMESTAMP: u64 = 1153704088;
+
 /// Allows writing the module to a wheel or add it directly to the virtualenv
 pub trait ModuleWriter {
-    /// Adds a directory relative to the module base path
-    fn add_directory(&mut self, path: impl AsRef<Path>) -> Result<()>;
-
     /// Adds a file with bytes as content in target relative to the module base path.
     ///
     /// For generated files, `source` is `None`.
@@ -179,13 +186,6 @@ impl PathWriter {
 }
 
 impl ModuleWriter for PathWriter {
-    fn add_directory(&mut self, path: impl AsRef<Path>) -> Result<()> {
-        let target = self.base_path.join(path);
-        debug!("Adding directory {}", target.display());
-        fs::create_dir_all(target)?;
-        Ok(())
-    }
-
     fn add_bytes_with_permissions(
         &mut self,
         target: impl AsRef<Path>,
@@ -198,6 +198,11 @@ impl ModuleWriter for PathWriter {
         if !self.file_tracker.add_file(target.as_ref(), source)? {
             // Ignore duplicate files.
             return Ok(());
+        }
+
+        if let Some(parent_dir) = path.parent() {
+            fs::create_dir_all(parent_dir)
+                .with_context(|| format!("Failed to create directory {}", parent_dir.display()))?;
         }
 
         // We only need to set the executable bit on unix
@@ -267,10 +272,6 @@ impl CompressionOptions {
 }
 
 impl ModuleWriter for WheelWriter {
-    fn add_directory(&mut self, _path: impl AsRef<Path>) -> Result<()> {
-        Ok(()) // We don't need to create directories in zip archives
-    }
-
     fn add_bytes_with_permissions(
         &mut self,
         target: impl AsRef<Path>,
@@ -432,13 +433,10 @@ pub struct SDistWriter {
     path: PathBuf,
     file_tracker: FileTracker,
     excludes: Override,
+    mtime: u64,
 }
 
 impl ModuleWriter for SDistWriter {
-    fn add_directory(&mut self, _path: impl AsRef<Path>) -> Result<()> {
-        Ok(())
-    }
-
     fn add_bytes_with_permissions(
         &mut self,
         target: impl AsRef<Path>,
@@ -446,6 +444,12 @@ impl ModuleWriter for SDistWriter {
         bytes: &[u8],
         permissions: u32,
     ) -> Result<()> {
+        if let Some(source) = source {
+            if self.exclude(source) {
+                return Ok(());
+            }
+        }
+
         let target = target.as_ref();
         if self.exclude(target) {
             return Ok(());
@@ -459,6 +463,7 @@ impl ModuleWriter for SDistWriter {
         let mut header = tar::Header::new_gnu();
         header.set_size(bytes.len() as u64);
         header.set_mode(permissions);
+        header.set_mtime(self.mtime);
         header.set_cksum();
         self.tar
             .append_data(&mut header, target, bytes)
@@ -466,28 +471,6 @@ impl ModuleWriter for SDistWriter {
                 "Failed to add {} bytes to sdist as {}",
                 bytes.len(),
                 target.display()
-            ))?;
-        Ok(())
-    }
-
-    fn add_file(&mut self, target: impl AsRef<Path>, source: impl AsRef<Path>) -> Result<()> {
-        let source = source.as_ref();
-        if self.exclude(source) {
-            return Ok(());
-        }
-        let target = target.as_ref();
-        if !self.file_tracker.add_file(target, Some(source))? {
-            // Ignore duplicate files.
-            return Ok(());
-        }
-
-        debug!("Adding {} from {}", target.display(), source.display());
-        self.tar
-            .append_path_with_name(source, target)
-            .context(format!(
-                "Failed to add file from {} to sdist as {}",
-                source.display(),
-                target.display(),
             ))?;
         Ok(())
     }
@@ -499,6 +482,7 @@ impl SDistWriter {
         wheel_dir: impl AsRef<Path>,
         metadata24: &Metadata24,
         excludes: Override,
+        mtime_override: Option<u64>,
     ) -> Result<Self, io::Error> {
         let path = wheel_dir
             .as_ref()
@@ -511,13 +495,15 @@ impl SDistWriter {
             .into_path_buf();
 
         let enc = GzEncoder::new(Vec::new(), Compression::default());
-        let tar = tar::Builder::new(enc);
+        let mut tar = tar::Builder::new(enc);
+        tar.mode(tar::HeaderMode::Deterministic);
 
         Ok(Self {
             tar,
             path,
             file_tracker: FileTracker::default(),
             excludes,
+            mtime: mtime_override.unwrap_or(SDIST_DETERMINISTIC_TIMESTAMP),
         })
     }
 
@@ -874,7 +860,11 @@ pub fn write_bindings_module(
     let ext_name = &project_layout.extension_name;
     let so_filename = if is_abi3 {
         if target.is_unix() {
-            format!("{ext_name}.abi3.so")
+            if target.is_cygwin() {
+                format!("{ext_name}.abi3.dll")
+            } else {
+                format!("{ext_name}.abi3.so")
+            }
         } else {
             match python_interpreter {
                 Some(python_interpreter) if python_interpreter.is_windows_debug() => {
@@ -939,7 +929,6 @@ pub fn write_bindings_module(
         }
     } else {
         let module = PathBuf::from(ext_name);
-        writer.add_directory(&module)?;
         // Reexport the shared library as if it were the top level module
         writer.add_bytes(
             module.join("__init__.py"),
@@ -1016,12 +1005,8 @@ pub fn write_cffi_module(
             .strip_prefix(python_module.parent().unwrap())
             .unwrap();
         module = relative.join(&project_layout.extension_name);
-        if !editable {
-            writer.add_directory(&module)?;
-        }
     } else {
         module = PathBuf::from(module_name);
-        writer.add_directory(&module)?;
         let type_stub = project_layout
             .rust_module
             .join(format!("{module_name}.pyi"));
@@ -1134,21 +1119,6 @@ fn generate_uniffi_bindings(
         .into_path_buf();
     fs::create_dir_all(&binding_dir)?;
 
-    let pattern = crate_dir.join("src").join("*.udl");
-    let udls = glob::glob(pattern.to_str().unwrap())?
-        .map(|p| p.unwrap())
-        .collect::<Vec<_>>();
-    let is_library = if udls.is_empty() {
-        true
-    } else if udls.len() > 1 {
-        bail!(
-            "Multiple UDL files found in {}",
-            crate_dir.join("src").display()
-        );
-    } else {
-        false
-    };
-
     let mut cmd = uniffi_bindgen_command(crate_dir)?;
     cmd.args([
         "generate",
@@ -1167,25 +1137,17 @@ fn generate_uniffi_bindings(
             .bindings
             .get("python")
             .and_then(|py| py.cdylib_name.clone());
-        if !is_library {
-            cmd.arg("--config");
-            cmd.arg(config_file);
-        }
+
+        // TODO: is this needed? `uniffi-bindgen` uses `uniffi.toml` by default,
+        // `uniffi_bindgen_command` sets cwd to the crate (workspace) root, so maybe
+        // we don't need to pass the config file explicitly?
+        cmd.arg("--config");
+        cmd.arg(config_file);
     }
 
-    let py_binding_name = if is_library {
-        cmd.arg("--library");
-        cmd.arg(artifact);
-        let file_stem = artifact.file_stem().unwrap().to_str().unwrap();
-        file_stem
-            .strip_prefix("lib")
-            .unwrap_or(file_stem)
-            .to_string()
-    } else {
-        let udl = &udls[0];
-        cmd.arg(udl);
-        udl.file_stem().unwrap().to_str().unwrap().to_string()
-    };
+    cmd.arg("--library");
+    cmd.arg(artifact);
+
     debug!("Running {:?}", cmd);
     let mut child = cmd.spawn().context(
         "Failed to run uniffi-bindgen, did you install it? Try `pip install uniffi-bindgen`",
@@ -1195,15 +1157,16 @@ fn generate_uniffi_bindings(
         bail!("Command {:?} failed", cmd);
     }
 
-    // uniffi bindings hardcoded the extension filenames
-    let cdylib_name = match cdylib_name {
-        Some(name) => name,
-        None => format!("uniffi_{py_binding_name}"),
-    };
-    let cdylib = match target_os {
-        Os::Macos => format!("lib{cdylib_name}.dylib"),
-        Os::Windows => format!("{cdylib_name}.dll"),
-        _ => format!("lib{cdylib_name}.so"),
+    // Name of the cdylib is either from uniffi.toml or derived from the library
+    let cdylib = match cdylib_name {
+        // this logic should match with uniffi's expected names, e.g.
+        // https://github.com/mozilla/uniffi-rs/blob/86a34083dd18bdd33f420c602b4fad624cc1e404/uniffi_bindgen/src/bindings/python/templates/NamespaceLibraryTemplate.py#L14-L37
+        Some(cdylib_name) => match target_os {
+            Os::Macos => format!("lib{cdylib_name}.dylib"),
+            Os::Windows => format!("{cdylib_name}.dll"),
+            _ => format!("lib{cdylib_name}.so"),
+        },
+        None => artifact.file_name().unwrap().to_str().unwrap().to_string(),
     };
 
     let py_bindings = fs::read_dir(&binding_dir)?
@@ -1283,12 +1246,8 @@ pub fn write_uniffi_module(
             .strip_prefix(python_module.parent().unwrap())
             .unwrap();
         module = relative.join(&project_layout.extension_name);
-        if !editable {
-            writer.add_directory(&module)?;
-        }
     } else {
         module = PathBuf::from(module_name);
-        writer.add_directory(&module)?;
         let type_stub = project_layout
             .rust_module
             .join(format!("{module_name}.pyi"));
@@ -1326,8 +1285,6 @@ pub fn write_bin(
         &metadata.version
     ))
     .join("scripts");
-
-    writer.add_directory(&data_dir)?;
 
     // We can't use add_file since we need to mark the file as executable
     writer.add_file_with_permissions(data_dir.join(bin_name), artifact, 0o755)?;
@@ -1420,9 +1377,7 @@ pub fn write_python_part(
                 continue;
             }
             let relative = absolute.strip_prefix(python_dir).unwrap();
-            if absolute.is_dir() {
-                writer.add_directory(relative)?;
-            } else {
+            if !absolute.is_dir() {
                 // Ignore native libraries from develop, if any
                 if let Some(extension) = relative.extension() {
                     if extension.to_string_lossy() == "so" {
@@ -1456,9 +1411,7 @@ pub fn write_python_part(
                     .filter_map(Result::ok)
                 {
                     let target = source.strip_prefix(pyproject_dir)?.to_path_buf();
-                    if source.is_dir() {
-                        writer.add_directory(target)?;
-                    } else {
+                    if !source.is_dir() {
                         #[cfg(unix)]
                         let mode = source.metadata()?.permissions().mode();
                         #[cfg(not(unix))]
@@ -1481,8 +1434,6 @@ pub fn write_dist_info(
     tags: &[String],
 ) -> Result<()> {
     let dist_info_dir = metadata24.get_dist_info_dir();
-
-    writer.add_directory(&dist_info_dir)?;
 
     writer.add_bytes(
         dist_info_dir.join("METADATA"),
@@ -1516,7 +1467,6 @@ pub fn write_dist_info(
 
     if !metadata24.license_files.is_empty() {
         let license_files_dir = dist_info_dir.join("licenses");
-        writer.add_directory(&license_files_dir)?;
         for path in &metadata24.license_files {
             writer.add_file(license_files_dir.join(path), pyproject_dir.join(path))?;
         }
@@ -1578,7 +1528,7 @@ pub fn add_data(
                     } else if file.path().is_file() {
                         writer.add_file_with_permissions(relative, file.path(), mode)?;
                     } else if file.path().is_dir() {
-                        writer.add_directory(relative)?;
+                        // Intentionally ignored
                     } else {
                         bail!("Can't handle data dir entry {}", file.path().display());
                     }
@@ -1606,7 +1556,7 @@ mod tests {
 
         // No excludes
         let tmp_dir = TempDir::new()?;
-        let mut writer = SDistWriter::new(&tmp_dir, &metadata, Override::empty())?;
+        let mut writer = SDistWriter::new(&tmp_dir, &metadata, Override::empty(), None)?;
         assert!(writer.file_tracker.0.is_empty());
         writer.add_bytes_with_permissions("test", Some(Path::new("test")), &[], perm)?;
         assert_eq!(writer.file_tracker.0.len(), 1);
@@ -1618,7 +1568,7 @@ mod tests {
         let mut excludes = OverrideBuilder::new(&tmp_dir);
         excludes.add("test*")?;
         excludes.add("!test2")?;
-        let mut writer = SDistWriter::new(&tmp_dir, &metadata, excludes.build()?)?;
+        let mut writer = SDistWriter::new(&tmp_dir, &metadata, excludes.build()?, None)?;
         writer.add_bytes_with_permissions("test1", Some(Path::new("test1")), &[], perm)?;
         writer.add_bytes_with_permissions("test3", Some(Path::new("test3")), &[], perm)?;
         assert!(writer.file_tracker.0.is_empty());