+++ src/python_interpreter/mod.rs
@@ -292,6 +292,8 @@ fn fun_with_abiflags(
     if bridge != &BridgeModel::Cffi
         && target.get_python_os() != message.system
         && !target.cross_compiling()
+        && !(target.get_python_os() == "cygwin"
+            && message.system.to_lowercase().starts_with("cygwin"))
     {
         bail!(
             "platform.system() in python, {}, and the rust target, {:?}, don't match ಠ_ಠ",
@@ -407,7 +409,7 @@ impl PythonInterpreter {
                 }
                 InterpreterKind::PyPy => {
                     // pypy uses its version as part of the ABI, e.g.
-                    // pypy 3.7 7.3 => numpy-1.20.1-pp37-pypy37_pp73-manylinux2014_x86_64.whl
+                    // pypy 3.11 7.3 => numpy-1.20.1-pp311-pypy311_pp73-manylinux2014_x86_64.whl
                     format!(
                         "pp{major}{minor}-{abi_tag}-{platform}",
                         major = self.major,